# perf: use DocumentEditor in code fixes