--- a/Chickensoft.AutoInject.Analyzers/src/utils/MethodModifier.cs
+++ b/Chickensoft.AutoInject.Analyzers/src/utils/MethodModifier.cs
@@ -1,48 +1,68 @@
 namespace Chickensoft.AutoInject.Analyzers.Utils;
 
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Microsoft.CodeAnalysis.Formatting;
+using Microsoft.CodeAnalysis.Editing;
 
 public static class MethodModifier {
   /// <summary>
-  /// Adds a this.method call to the end of a specified method within a type
-  /// declaration.
+  /// Creates a this.method call expression syntax node for the given method
+  /// name.
   /// </summary>
-  /// <param name="document">Document to modify.</param>
-  /// <param name="typeDeclaration">Type declaration off the class.</param>
-  /// <param name="originalMethodNode">The method to add the call to.</param>
   /// <param name="methodToCallName">
-  /// Name of the method to call at the end of the target method.
+  /// The name of the member method to call.
   /// </param>
-  /// <param name="cancellationToken">Cancellation Token</param>
-  /// <returns>Modified document</returns>
-  public static async Task<Document> AddCallToMethod(
-    Document document,
-    MethodDeclarationSyntax originalMethodNode,
+  /// <returns>
+  /// An expression node representing a "this.method()" expression.
+  /// </returns>
+  public static InvocationExpressionSyntax ThisMemberCallExpression(
     string methodToCallName,
-    CancellationToken cancellationToken
+    IEnumerable<string> argumentNames
   ) {
-    // Construct the parameterless call statement
-    var parameterlessCallStatement = SyntaxFactory.ExpressionStatement(
-      SyntaxFactory.InvocationExpression(
+    var invocation = SyntaxFactory.InvocationExpression(
         SyntaxFactory.MemberAccessExpression(
           SyntaxKind.SimpleMemberAccessExpression,
           SyntaxFactory.ThisExpression(),
-          SyntaxFactory.IdentifierName(methodToCallName)))
-    ).WithAdditionalAnnotations(Formatter.Annotation);
-
-    // Delegate to the more general helper
-    return await AddStatementToMethodBodyAsync(
-      document,
-      originalMethodNode,
-      parameterlessCallStatement,
-      cancellationToken
+          SyntaxFactory.IdentifierName(methodToCallName)
+        )
     );
+    if (argumentNames.Any()) {
+      invocation = invocation.WithArgumentList(
+        SyntaxFactory.ArgumentList(
+          SyntaxFactory.SeparatedList(
+            argumentNames.Select(
+              name =>
+                SyntaxFactory.Argument(SyntaxFactory.IdentifierName(name))
+            )
+          )
+        )
+      );
+    }
+    return invocation;
   }
+
+  /// <summary>
+  /// Creates a this.method call statement syntax node for the given method
+  /// name.
+  /// </summary>
+  /// <param name="methodToCallName">
+  /// The name of the member method to call.
+  /// </param>
+  /// <returns>
+  /// An expression statement node representing a "this.method()" statement.
+  /// </returns>
+  public static ExpressionStatementSyntax ThisMemberCallStatement(
+    string methodToCallName,
+    IEnumerable<string> argumentNames
+  ) =>
+    SyntaxFactory.ExpressionStatement(
+      ThisMemberCallExpression(methodToCallName, argumentNames)
+    );
 
   /// <summary>
   /// Adds a provided statement to the end of a specified method's body within a
@@ -62,31 +82,16 @@ public static class MethodModifier {
     StatementSyntax statementToAdd,
     CancellationToken cancellationToken
   ) {
-    var root = await document
-      .GetSyntaxRootAsync(cancellationToken)
-      .ConfigureAwait(false);
+    var editor = await DocumentEditor.CreateAsync(document, cancellationToken);
 
-    if (root is null) {
-      return document;
-    }
-
-    var methodInProgress = originalMethodNode;
-    BlockSyntax finalNewBody;
-
+    List<SyntaxNode> statements = [];
     if (originalMethodNode.Body is not null) {
-      // Existing block body
-      var existingStatements = originalMethodNode.Body.Statements;
-      var updatedStatements = existingStatements.Add(statementToAdd);
-      finalNewBody = originalMethodNode.Body.WithStatements(updatedStatements);
+      statements.AddRange(editor.Generator.GetStatements(originalMethodNode));
     }
     else {
-      // Expression body or missing body
-      var statementsForNewBlock = SyntaxFactory.List<StatementSyntax>();
-      if (originalMethodNode.ExpressionBody is not null) {
-        var originalExpression = originalMethodNode.ExpressionBody.Expression;
-        var statementFromExpr = SyntaxFactory
-          .ExpressionStatement(originalExpression);
-
+      var exprBodyExpr = editor.Generator.GetExpression(originalMethodNode) as ExpressionSyntax;
+      if (exprBodyExpr is not null) {
+        var statementFromExpr = SyntaxFactory.ExpressionStatement(exprBodyExpr);
         // Make sure to preserve the trailing trivia from the original method's semicolon token
         // If we don't do this we will lose any code comments or whitespace that was after the semicolon
         var originalMethodSemicolon = originalMethodNode.SemicolonToken;
@@ -105,27 +110,13 @@ public static class MethodModifier {
               );
           }
         }
-        statementsForNewBlock = statementsForNewBlock.Add(statementFromExpr);
-
-        // Remove the old expression body from the method
-        methodInProgress = methodInProgress
-            .WithExpressionBody(null)
-            .WithSemicolonToken(SyntaxFactory.Token(SyntaxKind.None));
+        statements.Add(statementFromExpr);
       }
-
-      // Add the provided statement
-      statementsForNewBlock = statementsForNewBlock.Add(statementToAdd);
-      // Create a new block with the collected statements
-      finalNewBody = SyntaxFactory
-        .Block(statementsForNewBlock);
     }
+    statements.Add(statementToAdd);
 
-    var fullyModifiedMethod = methodInProgress
-      .WithBody(finalNewBody)
-      .WithAdditionalAnnotations(Formatter.Annotation);
+    editor.SetStatements(originalMethodNode, statements);
 
-    // Replace the original method with the modified one in the type declaration
-    var newRoot = root.ReplaceNode(originalMethodNode, fullyModifiedMethod);
-    return document.WithSyntaxRoot(newRoot);
+    return editor.GetChangedDocument();
   }
 }