Using the DocumentEditor class instead of modifying syntax trees directly, as well as caching static syntax trees where possible, results in mild performance increases for the code fixes.

## Current Benchmarks (Before Change)
| Type                               | Method                            | Mean          | Error       | StdDev      | Median        | Ratio    | RatioSD  | Gen0      | Gen1      | Allocated    | Alloc Ratio |
|----------------------------------- |---------------------------------- |--------------:|------------:|------------:|--------------:|---------:|---------:|----------:|----------:|-------------:|------------:|
| NotificationFixBenchmark           | NotificationFix                   | 211,346.85 μs | 1,380.56 μs | 1,291.38 μs | 211,406.00 μs | 4,640.29 | 2,038.41 | 8000.0000 | 1000.0000 | 135001.52 KB |    4,285.76 |
| NotifyFixBenchmark                 | NotifyFix                         | 194,434.23 μs | 1,654.77 μs | 1,381.81 μs | 194,266.05 μs | 4,268.96 | 1,875.45 | 7000.0000 | 1000.0000 | 128973.48 KB |    4,094.40 |
| ProvideFixExistingMethodsBenchmark | ProvideFixExistingSetup           | 199,957.55 μs | 1,932.50 μs | 1,613.72 μs | 199,839.80 μs | 4,390.23 | 1,928.81 | 7000.0000 |         - | 129484.88 KB |    4,110.63 |
| ProvideFixNewMethodsBenchmark      | ProvideFixNewSetup                | 221,748.94 μs | 1,232.16 μs | 1,092.28 μs | 221,620.45 μs | 4,868.68 | 2,138.70 | 8000.0000 | 2000.0000 |  136877.8 KB |    4,345.33 |
| ProvideFixExistingMethodsBenchmark | ProvideFixExistingOnReady         | 195,541.62 μs |   921.38 μs |   719.35 μs | 195,748.05 μs | 4,293.27 | 1,885.99 | 7000.0000 | 1000.0000 |  129312.8 KB |    4,105.17 |
| ProvideFixNewMethodsBenchmark      | ProvideFixNewOnReady              | 220,897.49 μs | 1,955.46 μs | 1,733.46 μs | 220,628.85 μs | 4,849.98 | 2,130.72 | 8000.0000 | 2000.0000 | 136821.12 KB |    4,343.53 |
| ProvideFixExistingMethodsBenchmark | ProvideFixExistingReady           | 199,044.19 μs |   925.90 μs |   773.17 μs | 198,841.10 μs | 4,370.18 | 1,919.72 | 7000.0000 | 1000.0000 | 129275.34 KB |    4,103.98 |
| ProvideFixNewMethodsBenchmark      | ProvideFixNewReady                | 225,705.29 μs | 2,740.69 μs | 2,429.55 μs | 225,632.00 μs | 4,955.54 | 2,177.43 | 8000.0000 | 1000.0000 | 137130.23 KB |    4,353.34 |

## Updated Benchmarks (After Change)
| Type                               | Method                            | Mean          | Error        | StdDev      | Median        | Ratio    | RatioSD  | Gen0      | Gen1      | Allocated    | Alloc Ratio |
|----------------------------------- |---------------------------------- |--------------:|-------------:|------------:|--------------:|---------:|---------:|----------:|----------:|-------------:|------------:|
| NotificationFixBenchmark           | NotificationFix                   | 160,486.86 μs | 1,821.195 μs | 1,614.44 μs | 160,296.40 μs | 3,867.60 | 1,728.52 | 6000.0000 | 2000.0000 | 110707.91 KB |    3,514.54 |
| NotifyFixBenchmark                 | NotifyFix                         | 179,470.32 μs | 1,065.618 μs |   996.78 μs | 179,461.60 μs | 4,325.09 | 1,932.56 | 7000.0000 | 1000.0000 | 114999.85 KB |    3,650.79 |
| ProvideFixExistingMethodsBenchmark | ProvideFixExistingSetup           | 186,992.67 μs | 1,543.031 μs | 1,288.50 μs | 186,717.60 μs | 4,506.37 | 2,013.75 | 7000.0000 | 1000.0000 | 115910.01 KB |    3,679.68 |
| ProvideFixNewMethodsBenchmark      | ProvideFixNewSetup                | 178,469.68 μs | 1,364.425 μs | 1,209.53 μs | 178,125.70 μs | 4,300.98 | 1,921.91 | 7000.0000 | 2000.0000 | 114715.11 KB |    3,641.75 |
| ProvideFixExistingMethodsBenchmark | ProvideFixExistingOnReady         | 181,349.88 μs | 1,964.811 μs | 1,837.89 μs | 181,340.80 μs | 4,370.39 | 1,953.19 | 7000.0000 | 1000.0000 | 115617.56 KB |    3,670.40 |
| ProvideFixNewMethodsBenchmark      | ProvideFixNewOnReady              | 177,024.35 μs | 1,424.504 μs | 1,189.53 μs | 177,604.20 μs | 4,266.14 | 1,906.39 | 7000.0000 | 2000.0000 | 114672.21 KB |    3,640.39 |
| ProvideFixExistingMethodsBenchmark | ProvideFixExistingReady           | 187,872.30 μs | 1,361.200 μs | 1,062.74 μs | 187,803.75 μs | 4,527.57 | 2,023.20 | 7000.0000 | 1000.0000 | 115890.62 KB |    3,679.07 |
| ProvideFixNewMethodsBenchmark      | ProvideFixNewReady                | 177,974.32 μs | 1,200.186 μs | 1,002.21 μs | 178,021.20 μs | 4,289.04 | 1,916.55 | 7000.0000 | 1000.0000 |  115226.5 KB |    3,657.98 |